+++ src/resolvers/MultisigResolver.sol
@@ -28,12 +28,12 @@ contract MultisigResolver {
     constructor(address _builderEngine, address[] memory _signers, uint256 _threshold) {
         require(_signers.length >= _threshold, "InvalidThreshold");
         require(_threshold > 0, "ZeroThreshold");
-        
+
         builderEngine = _builderEngine;
         signers = _signers;
         threshold = _threshold;
-        
-        for (uint i = 0; i < _signers.length; i++) {
+
+        for (uint256 i = 0; i < _signers.length; i++) {
             isSigner[_signers[i]] = true;
         }
     }
