+++ test/Resolution.t.sol
@@ -8,12 +8,13 @@ contract ResolutionTest is BaseTest {
         super.setUp();
         vm.prank(alice);
         covenantJoin.joinCovenant{value: 0.01 ether}();
-        
+
         vm.prank(alice);
         builderEngine.submitProposal("Job", 0.005 ether);
 
         // Fast forward to Funded
-        vm.prank(highRep1); builderEngine.approveProposal(0);
+        vm.prank(highRep1);
+        builderEngine.approveProposal(0);
 
         // Fast forward to Delivered
         vm.prank(alice);
@@ -28,19 +29,19 @@ contract ResolutionTest is BaseTest {
 
     function testResolveSuccessPayoutAndUnlock() public {
         uint256 preBal = alice.balance;
-        
+
         vm.prank(admin);
         builderEngine.resolveProposal(0, true);
 
         // Payout to Alice
         assertEq(alice.balance, preBal + 0.005 ether);
-        
+
         // Lock released
         assertEq(builderEngine.lockedFunds(), 0);
-        
+
         // Status Completed
         (,,,, BuilderEngine.Status status,,) = builderEngine.proposals(0);
-        assertEq(uint(status), uint(BuilderEngine.Status.Completed));
+        assertEq(uint256(status), uint256(BuilderEngine.Status.Completed));
     }
 
     function testResolveFailRefundAndUnlock() public {
@@ -52,13 +53,13 @@ contract ResolutionTest is BaseTest {
 
         // No payout
         assertEq(alice.balance, preBal);
-        
+
         // Contract keeps funds (released from lock)
         assertEq(address(builderEngine).balance, contractBal);
         assertEq(builderEngine.lockedFunds(), 0);
 
         // Status Failed
         (,,,, BuilderEngine.Status status,,) = builderEngine.proposals(0);
-        assertEq(uint(status), uint(BuilderEngine.Status.Failed));
+        assertEq(uint256(status), uint256(BuilderEngine.Status.Failed));
     }
 }
